chore: move enums into relevant modules
