+++ benchmarking/data/real_data/cosc341_w2021_t2_provider/get_statistics.py
@@ -1,4 +1,4 @@
-from api.dataclasses.enums import ScenarioAttribute, Gender, YearLevel
+from benchmarking.evaluations.enums import ScenarioAttribute, Gender, YearLevel
 from benchmarking.data.real_data.cosc341_w2021_t2_provider.providers import (
     COSC341W2021T2StudentProvider,
 )
