+++ benchmarking/evaluations/scenarios/concentrate_timeslots_and_concentrate_gender_and_concentrate_race.py
@@ -2,6 +2,8 @@ from typing import List
 
 from api.dataclasses.enums import (
     DiversifyType,
+)
+from benchmarking.evaluations.enums import (
     ScenarioAttribute,
     Gender,
     Race,
