+++ benchmarking/evaluations/scenarios/cosc341/concentrate_timeslot_diversify_gender_min_2_and_diversify_year_level.py
@@ -2,9 +2,11 @@ from typing import List
 
 from api.dataclasses.enums import (
     DiversifyType,
+    TokenizationConstraintDirection,
+)
+from benchmarking.evaluations.enums import (
     ScenarioAttribute,
     Gender,
-    TokenizationConstraintDirection,
 )
 from api.dataclasses.tokenization_constraint import TokenizationConstraint
 from benchmarking.evaluations.goals import WeightGoal, DiversityGoal
