+++ benchmarking/evaluations/scenarios/diversify_gender_only.py
@@ -1,6 +1,7 @@
 from typing import List
 
-from api.dataclasses.enums import DiversifyType, ScenarioAttribute
+from api.dataclasses.enums import DiversifyType
+from benchmarking.evaluations.enums import ScenarioAttribute
 from benchmarking.evaluations.interfaces import Scenario, Goal
 from benchmarking.evaluations.goals import (
     DiversityGoal,
