+++ benchmarking/evaluations/scenarios/project_location.py
@@ -1,6 +1,7 @@
 from typing import List
 
-from api.dataclasses.enums import DiversifyType, ScenarioAttribute, RequirementsCriteria
+from api.dataclasses.enums import DiversifyType
+from benchmarking.evaluations.enums import ScenarioAttribute
 from benchmarking.evaluations.goals import (
     DiversityGoal,
     WeightGoal,
