+++ benchmarking/evaluations/scenarios/project_requirement_preferences_diversify_gender.py
@@ -1,7 +1,7 @@
 from typing import List
 
-from api.dataclasses.enums import DiversifyType, ScenarioAttribute, RequirementsCriteria
-from benchmarking.evaluations.enums import PreferenceDirection, PreferenceSubject
+from api.dataclasses.enums import DiversifyType, PreferenceDirection, PreferenceSubject
+from benchmarking.evaluations.enums import ScenarioAttribute
 from benchmarking.evaluations.goals import (
     DiversityGoal,
     WeightGoal,
