+++ benchmarking/evaluations/scenarios/satisfy_project_requirements_and_diversify_female_min_of_2_and_diversify_african_min_of_2.py
@@ -1,10 +1,11 @@
 from typing import List
 
 from api.dataclasses.enums import (
-    RequirementsCriteria,
     DiversifyType,
-    ScenarioAttribute,
     TokenizationConstraintDirection,
+)
+from benchmarking.evaluations.enums import (
+    ScenarioAttribute,
     Gender,
     Race,
 )
