+++ benchmarking/runs/diversify_gender_min_2.py
@@ -11,7 +11,8 @@ from api.ai.interfaces.algorithm_config import (
 )
 from api.ai.priority_algorithm.mutations.local_max import LocalMaxMutation
 from api.ai.priority_algorithm.mutations.random_swap import RandomSwapMutation
-from api.dataclasses.enums import ScenarioAttribute, Gender, AlgorithmType
+from api.dataclasses.enums import AlgorithmType
+from benchmarking.evaluations.enums import ScenarioAttribute, Gender
 from benchmarking.data.simulated_data.mock_student_provider import (
     MockStudentProvider,
     MockStudentProviderSettings,
