+++ benchmarking/runs/priority_algorithm/internal_parameter_exploration/projects_run.py
@@ -9,12 +9,11 @@ from api.ai.interfaces.algorithm_config import (
     PriorityAlgorithmConfig,
     PriorityAlgorithmStartType,
 )
-from api.dataclasses.enums import RequirementsCriteria, Gpa
 from api.dataclasses.enums import (
-    ScenarioAttribute,
     RequirementOperator,
     AlgorithmType,
 )
+from benchmarking.evaluations.enums import ScenarioAttribute, Gpa
 from api.dataclasses.project import Project, ProjectRequirement
 from benchmarking.data.simulated_data.mock_initial_teams_provider import (
     MockInitialTeamsProvider,
