+++ benchmarking/runs/real_data/cosc_341/concentrate_time_gender_diversify_yl_team_size_6.py
@@ -9,7 +9,8 @@ from api.ai.interfaces.algorithm_config import (
     WeightAlgorithmConfig,
     GroupMatcherAlgorithmConfig,
 )
-from api.dataclasses.enums import ScenarioAttribute, Gender, AlgorithmType
+from api.dataclasses.enums import AlgorithmType
+from benchmarking.evaluations.enums import ScenarioAttribute, Gender
 from benchmarking.data.real_data.cosc341_w2021_t2_provider.providers import (
     COSC341W2021T2AnsweredSurveysStudentProvider,
 )
